Improved separation of concerns
- Improved the separation of concerns by moving the implementation of services out of the Framework, and into the App layer.
- Sources/AllToolsMenu → DittoToolsApp/MenuView
- Refactored ToolsList → AllToolsMenu
- removed unused classes, stale code
