+++ DittoToolsApp/DittoToolsApp/Assets.xcassets/Ditto.LogoMark.Blue.imageset/Contents.json
