+++ DittoToolsApp/DittoToolsApp/Model/DittoIdentity+Extension.swift
