+++ DittoToolsApp/DittoToolsApp/Model/IdentityConfiguration.swift
