+++ DittoToolsApp/DittoToolsApp/Model/SupplementaryCredentials.swift
