+++ DittoToolsApp/DittoToolsApp/Services/AuthenticationDelegate.swift
