+++ DittoToolsApp/DittoToolsApp/Services/Ditto Service/DittoServiceError.swift
