+++ DittoToolsApp/DittoToolsApp/Views/Identity Configuration View/Components/ClearableTextField.swift
