+++ DittoToolsApp/DittoToolsApp/Views/Identity Configuration View/Identity Configuration Form/IdentityFormViewModel.swift
